Added FLA dataset files with randomization. Rerunning experiments with two scenes
Former-commit-id: 23460c7b719807b8e9789c22574db03aba49e373 [formerly b78938f98a4b8d13b8de8a90438c38da5a29b3d7]
Former-commit-id: f3b1a5663358e017827afd660887bbb4ac742736

+++ loaders.py
@@ -384,7 +384,7 @@ class FLADataset(tud.Dataset):
             lines = [line for line in lines if line[0] is not "#"] # Strip comments.
             for line in lines:
                 tokens = line.split(",")
-                self.image_pair_ids.append([np.uint64(tokens[0]), tokens[1]])
+                self.image_pair_ids.append([np.uint64(tokens[0]), np.uint64(tokens[1])])
         print('Loaded {} pairs of images from {}'.format(len(self.image_pair_ids), dataset_file))
        
     def __len__(self):
